Renaming
RelayHubApi -> IRelayHub
RelayRecipientApi -> IRelayRecipient

Updating server version
