+++ build_server_docker.sh
@@ -12,7 +12,7 @@ cp ./build/dock-builD/server/bin/RelayHttpServer $serverbuild
 cp ./serverdock/start-relay.sh $serverbuild
 cp ./serverdock/start-relay-with-ganache.sh $serverbuild
 cp ./scripts/fundrelay.js $serverbuild
-cp ./src/js/relayclient/RelayHubApi.js $serverbuild
+cp ./src/js/relayclient/IRelayHub.js $serverbuild
 cp ./serverdock/truffle.js $serverbuild
 cp -a ./contracts $serverbuild
 cp -a ./serverdock/migrations $serverbuild
