+++ server/src/RelayHttpServer.go
@@ -19,7 +19,7 @@ import (
 	"time"
 )
 
-const VERSION = "0.3.6"
+const VERSION = "0.4.0"
 
 var KeystoreDir = filepath.Join(os.Getenv("PWD"), "build/server/keystore")
 var delayBetweenRegistrations = 24 * int64(time.Hour/time.Second) // time.Duration is in nanosec - converting to sec like unix
