+++ src/js/webtools/contractmanager.js
@@ -1,11 +1,11 @@
 /* global web3 */
 const utils = require('./utils')
 const promisify = utils.promisify
-const RelayHubApi = require('../relayclient/RelayHubApi')
-const RelayRecipientApi = require('../relayclient/RelayRecipientApi')
+const IRelayHub = require('../relayclient/IRelayHub')
+const IRelayRecipient = require('../relayclient/IRelayRecipient')
 
-const RelayHub = web3.eth.contract(RelayHubApi)
-const RelayRecipient = web3.eth.contract(RelayRecipientApi)
+const RelayHub = web3.eth.contract(IRelayHub)
+const RelayRecipient = web3.eth.contract(IRelayRecipient)
 
 class ContractManager {
